requirements.txt
ffmpeg_python==0.2.0
ipython==8.12.3
langchain_community==0.2.5
python-dotenv==1.0.1
Requests==2.32.3
